Final commit - Project completed
Notes now save whenever anything changes, and data is loaded when app is rerun.

Done for now, but might add an extra feature or two later on.

Thinking about maybe making this sync with iCloud and create a watch app. That way users can take notes on both devices that are synced across the board.